[Rename] Create & refactor packages
