+++ src/main/java/com/ormi5/movieblog/post/Post.java
@@ -1,4 +1,4 @@
-package com.ormi5.movieblog;
+package com.ormi5.movieblog.post;
 
 import jakarta.persistence.*;
 import lombok.Getter;
