+++ src/main/java/com/ormi5/movieblog/postcontroller/PostController.java
@@ -1,5 +1,6 @@
-package com.ormi5.movieblog;
+package com.ormi5.movieblog.postcontroller;
 
+import com.ormi5.movieblog.post.PostDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -9,7 +10,6 @@ import java.util.Optional;
 @RestController
 @RequestMapping("/posts")
 public class PostController {
-
     private final PostService postService;
 
     @Autowired
