+++ src/main/java/com/ormi5/movieblog/postcontroller/PostRepository.java
@@ -1,5 +1,6 @@
-package com.ormi5.movieblog;
+package com.ormi5.movieblog.postcontroller;
 
+import com.ormi5.movieblog.post.Post;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
